refactor: Remove duplicated function on components
